Convert to VC2015 Express
Allowed VC2015 to convert the project files to VC2015. In this version
we are able to use incremental linking in debug but not in release. To
get this to work we need to us the VC7 versions of the CodaRTnetSDK
libraries.

+++ CodaRTnetAlignment/CodaRTnetAlignment.cpp
@@ -246,7 +246,7 @@ void perform_alignment( void ) {
 	cl.getDeviceInfo(info);
 
 	// print alignment diagnostics
-	DWORD marker_id_array[5] = { origin, x_negative, x_positive, xy_negative, xy_positive };
+	DWORD marker_id_array[5] = { (DWORD) origin, (DWORD) x_negative, (DWORD) x_positive, (DWORD) xy_negative, (DWORD) xy_positive };
 	print_alignment_status(marker_id_array, info);
 
 	// exit early if alignment failed
